refactor: move senders in to app/senders/
