+++ app/lib/message_dequeuer/incoming_message_processor.rb
@@ -163,11 +163,11 @@ module MessageDequeuer
 
       case queued_message.message.endpoint
       when SMTPEndpoint
-        sender = @state.sender_for(Postal::SMTPSender, queued_message.message.recipient_domain, nil, servers: [queued_message.message.endpoint])
+        sender = @state.sender_for(SMTPSender, queued_message.message.recipient_domain, nil, servers: [queued_message.message.endpoint])
       when HTTPEndpoint
-        sender = @state.sender_for(Postal::HTTPSender, queued_message.message.endpoint)
+        sender = @state.sender_for(HTTPSender, queued_message.message.endpoint)
       when AddressEndpoint
-        sender = @state.sender_for(Postal::SMTPSender, queued_message.message.endpoint.domain, nil, force_rcpt_to: queued_message.message.endpoint.address)
+        sender = @state.sender_for(SMTPSender, queued_message.message.endpoint.domain, nil, force_rcpt_to: queued_message.message.endpoint.address)
       else
         log "invalid endpoint for route (#{queued_message.message.endpoint_type})"
         create_delivery "HardFail", details: "Invalid endpoint for route."
