+++ app/senders/base_sender.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+class BaseSender
+
+  def start
+  end
+
+  def send_message(message)
+  end
+
+  def finish
+  end
+
+end
