--- lib/postal/sender.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module Postal
-  class Sender
-
-    def start
-    end
-
-    def send_message(message)
-    end
-
-    def finish
-    end
-
-  end
-end
